+++ src/main/java/com/jangbogo/mall/controller/AddressController.java
@@ -25,7 +25,10 @@ public class AddressController {
 
     // 배송지 페이지로 이동
     @GetMapping("/address/list")
-    public String product() {
+
+    public String product(HttpServletRequest req,Model m) {
+        m.addAttribute("mypageUrl", req.getRequestURI());
+
         return "/dlvpn/dlvpnlist";
     }
 
@@ -34,7 +37,7 @@ public class AddressController {
     public ResponseEntity<List<Address>> list(HttpSession session){
         List<Address> list = null;
         Integer user_idx = (Integer)session.getAttribute("idx"); // 세션에서 회원번호 가져오기
-        System.out.println("user_idx = " + user_idx);
+//        System.out.println("user_idx = " + user_idx);
         try {
             list = service.selAddrList(user_idx);
             return new ResponseEntity<List<Address>>(list, HttpStatus.OK);
@@ -152,87 +155,6 @@ public class AddressController {
 
     }
 
-
-/////유효성 검사 해야함 ! 번호 체크 !
-//    @PostMapping ("/addressremove") //배송지 수정창-> 삭제
-//    public String myPageAddrRemove(Address address,Integer idx,HttpServletRequest req, Model m, HttpSession session) {
-////        m.addAttribute("mypageUrl", req.getRequestURI());
-//        int user_idx = (int) session.getAttribute("idx"); //세션 얻기
-//
-//        // 회원번호랑, 배송지일련번호를 가져와서 삭제한다
-//        try {
-//            if(service.deleteAddr(idx,user_idx)!=1) throw new Exception("Delete failed.");
-//        } catch(Exception e) {
-//            e.printStackTrace();
-//        }
-//        return "redirect:/mypage/addresslist";
-////        return "/dlvpn/dlvpnlistfinal";
-//
-//    }
-////
-//    @PostMapping ("/addressupdate") //배송지 수정창-> 수정
-//    public String myPageAddrUpdate(Address address,HttpServletRequest req, Model m, HttpSession session) {
-//
-//
-//
-//        int user_idx = (int) session.getAttribute("idx");   // 세션에서 회원번호 회득
-//        address.setUser_idx(user_idx);
-//        try {
-//            if (address.getIs_default_yn().equals("true")) {
-//                service.resetDefault_N(user_idx);
-//                service.updateAddr(address);
-//            }else {
-//                service.updateAddr(address);
-//            }
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//        }
-//        return "redirect:/mypage/addresslist";
-////        return "/dlvpn/dlvpnlistfinal";
-//
-//    }
-
-
-
-
-
-    //
-//    @GetMapping("/addresslist") //배송지 조회
-//    public String myPageAddrList(HttpServletRequest req, Model m, HttpSession session,HttpServletRequest request) {
-//        m.addAttribute("mypageUrl", req.getRequestURI());
-//
-//        if(!loginCheck(request))
-//            return "redirect:/user/login?toURL="+request.getRequestURL();  // 로그인을 안했으면 로그인 화면으로 이동
-//
-//        int user_idx = (int) session.getAttribute("idx"); //세션에서 회원번호 얻기
-//
-//        try {
-//            List<Address> list = service.selAddrList(user_idx);
-//            m.addAttribute("addrList", list);
-//        } catch(Exception e) {
-//            e.printStackTrace();
-//        }
-//        return "dlvpn/dlvpnlist";
-//    }
-//
-
-//
-//    @PostMapping ("/changestate") // 배송지 선택 상태 변경
-//    public String chageState(Address address,HttpServletRequest req, Model m, HttpSession session) {
-//
-//        int user_idx = (int) session.getAttribute("idx");
-//        int idx = address.getIdx();
-//
-//        try {
-//            service.resetStateCD(user_idx);
-//            service.AddrStateCD(idx);
-//        } catch(Exception e) {
-//            e.printStackTrace();
-//        }
-//        return "redirect:/mypage/addresslist";
-//
-//    }
-//
 
     private boolean loginCheck(HttpServletRequest request) {
         // 1. 세션을 얻어서
