+++ src/main/webapp/resources/css/productReviewModal.css
@@ -88,6 +88,7 @@ display: block; */
     max-width: 100%;
     min-height: 100%;
     max-height: 100%; */
+    border-radius: 5px;
 }
 .img-name{
     font-size: 16px;
