+++ src/main/webapp/resources/css/wishlist.css
@@ -18,7 +18,7 @@ button {
     border-radius: 4px;
     color: rgb(51, 51, 51);
     background-color: rgb(255, 255, 255);
-    border: 1px solid rgb(221, 221, 221);
+    border: 1px solid navy;
     cursor: pointer;
     margin-bottom: 4px;
     /*width:104px ;*/
@@ -34,12 +34,18 @@ button {
     width: 104px;
     height: 36px;
     border-radius: 4px;
-    color: white;
-    background-color: #f58412;
-    border: 1px solid whitesmoke;
-    /*width:104px ;*/
-    /*height :36px;*/
-    /*radius: 4px;*/
+    color: #f58412;
+    background-color: white;
+    border: 1px solid #f58412;
+    font-weight: 500;
+
+}
+.openBtn img{
+    display: inline-block;
+    width: 20px;
+    height: 20px;
+    margin-right: 4px;
+    vertical-align: middle;
 }
 
 .css-102v0ri span {
@@ -110,7 +116,7 @@ a {
     text-decoration: line-through;
 }
 
-img {
+.wishlist-box a img {
     /* display: inline-block; */
     width: 60px;
     height: 78px;
